Encapsulate key/value variants for data transformations in base class
