--- connect/transforms/src/main/java/org/apache/kafka/connect/transforms/InsertInKey.java
@@ -1,44 +1,0 @@
-/**
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- * <p/>
- * http://www.apache.org/licenses/LICENSE-2.0
- * <p/>
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- **/
-
-package org.apache.kafka.connect.transforms;
-
-import org.apache.kafka.connect.connector.ConnectRecord;
-import org.apache.kafka.connect.data.Schema;
-
-/**
- * This transformation allows inserting configured attributes of the record metadata as fields in the record key.
- * It also allows adding a static data field.
- */
-public class InsertInKey<R extends ConnectRecord<R>> extends Insert<R> {
-
-    @Override
-    protected Schema operatingSchema(R record) {
-        return record.keySchema();
-    }
-
-    @Override
-    protected Object operatingValue(R record) {
-        return record.key();
-    }
-
-    @Override
-    protected R newRecord(R record, Schema updatedSchema, Object updatedValue) {
-        return record.newRecord(record.topic(), record.kafkaPartition(), updatedSchema, updatedValue, record.valueSchema(), record.value(), record.timestamp());
-    }
-
-}
